_site/requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autopep8==1.7.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.2
cryptography==38.0.1
Django==4.1.2
dnspython==2.2.1
email-validator==1.3.0
emoji==2.1.0
emojis==0.6.0
frozenlist==1.3.1
future==0.18.2
game==0.0.5
google==3.0.0
google-api-core==2.10.2
google-auth==2.14.0
google-cloud==0.34.0
google-cloud-speech==2.16.2
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
motor==3.0.0
multidict==6.0.2
proto-plus==1.22.1
protobuf==4.21.9
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pykerberos==1.2.4
pymediainfo==5.1.0
pymongo==4.3.2
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
python-multipart==0.0.5
python-snappy==0.6.1
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.21.0
toml==0.10.2
tornado==6.1
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uuid==1.30
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
webencodings==0.5.1
websockets==10.3
yarl==1.8.1
